update `xnft build|start|dev` to bundle into self contained html file
