+++ packages/xnft-cli/.gitignore
@@ -1,2 +1,3 @@
 renderer.js
 iframe.js
+iframe.html
