+++ packages/xnft-cli/legacy/build.js
@@ -1,0 +1,9 @@
+const esconfig = require("./esbuild.config.web");
+module.exports = (program) => {
+  program.command("build").action(() => {
+    require("esbuild").build({
+      ...esconfig,
+      minify: true,
+    });
+  });
+};
