+++ packages/xnft-cli/legacy/start.js
@@ -1,13 +1,12 @@
 const fs = require("fs");
 const express = require("express");
-const getHtmlWrapper = require("./getHtmlWrapper");
 const simulatorPort = require("../simulatorPort");
-module.exports = (program) => {
-  program.command("start").action(() => {
-    const app = express(); // create express app
 
+module.exports = (program) => {
+  program.command("start").action(async () => {
+    const app = express();
+    const port = simulatorPort;
     const html = fs.readFileSync("dist/index.html", { encoding: "utf-8" });
-    let port = simulatorPort;
 
     app.get("/", (req, res) => {
       res.send(html);
