+++ packages/xnft-cli/native/dev.js
@@ -1,18 +1,20 @@
 const fs = require("fs");
 const express = require("express");
 const esconfig = require("./esbuild.config.web");
+const simulatorPort = require("../simulatorPort");
 module.exports = (program) => {
   program.command("dev").action(() => {
     const app = express(); // create express app
 
     let html;
-    let port = 9933;
+    let port = simulatorPort;
 
     require("esbuild")
       .build({
         ...esconfig,
         watch: {
           onRebuild(error, result) {
+            console.log("build changed");
             if (error) console.error("build error", JSON.stringify(err));
             html = fs.readFileSync("dist/index.html", { encoding: "utf-8" });
           },
