+++ packages/xnft-cli/package.json
@@ -10,7 +10,7 @@
   ],
   "scripts": {
     "prebuild": "cd ../../examples/xnft/iframe && npm run build && cd ../../../packages/common && npm run build && cd ../react-xnft-dom-renderer && npm run build",
-    "build": "shx cp ../../examples/xnft/iframe/dist/index.js iframe.js && shx cp ../react-xnft-dom-renderer/dist/index.js renderer.js"
+    "build": "shx cp ../../examples/xnft/iframe/dist/index.html iframe.html && shx cp ../react-xnft-dom-renderer/dist/index.js renderer.js"
   },
   "browserslist": [
     "last 1 chrome version",
