+++ packages/xnft-cli/simulatorPort.js
@@ -1,0 +1,1 @@
+module.exports = 9933;
